Fixed Linux compilation errors

+++ Amicable/PrimeTables.cpp
@@ -253,7 +253,7 @@ NOINLINE void SearchCandidates(Factor* factors, const num64 value, const num64 s
 		const int index_inv128 = f.index - 1;
 		if (index_inv128 >= 0)
 		{
-			f.p_inv128 = (index_inv128 < ReciprocalsTableSize128) ? PrimeInverses128[index_inv128] : -modular_inverse128(f.p.Get());
+			f.p_inv128 = (static_cast<unsigned int>(index_inv128) < ReciprocalsTableSize128) ? PrimeInverses128[index_inv128] : -modular_inverse128(f.p.Get());
 		}
 
 		for (;;)
